// types/car-specs.ts
export enum ECarTypes {
  SUV = 'SUV',
  MINIVAN = 'Minivan',
  PICKUP_TRUCK = 'Pickup Truck',
  SPORTS_CAR = 'Sports Car',
  HATCHBACK = 'Hatchback',
  SEDAN = 'Sedan',
}

export enum ETransmissions {
  AUTOMATIC = 'Automatic',
  MANUAL = 'Manual',
}

export enum EEngineTypes {
  GAS = 'Gas',
  HYBRID = 'Hybrid',
  ELECTRIC = 'Electric',
}

export enum ESeats {
  ONE = '1',
  TWO = '2',
  THREE = '3',
  FOUR = '4',
  FIVE = '5',
  SIX = '6',
  SEVEN = '7',
}